Fixes 568
Set the x-amz-website-redirect-location property for the object that
represents the index.html file of the same named directory.

The deployment script already creates a same named object for any
directory that contains an index.html file. We now set the
x-amz-website-redirect-location property for this object so that access
to `/documentation/publish/submitting-an-add-on` for example, gets 301
redirected to `/documentation/publish/submitting-an-add-on/`.