Removal of unlimitedStorage prompting (#633)
Bug [1630413](https://bugzilla.mozilla.org/show_bug.cgi?id=1630413) amends the handling of `unlimitedStorage` so that, from Firefox 77, it is no longer prompted when an extension is installed or updated. Also, removed the hyphens preceding the permissions in the "displayed" table. On MDN these hyphens used to display as bullets. Also added a new subheading for  "Other permissions" to clarify the "displayed" and "not displayed" list content.

Co-authored-by: Richard Bloor <rbloor@atlassian.com>
Co-authored-by: Caitlin Neiman <cneiman@mozilla.com>