Add mention of --chromium-binary to --target documentation
When using `--target chromium`, the user can further specify which
Chromium browser they want to use with `--chromium-binary`.

This is not clear from current documentation and comes up rather often
so decided to document here.

Also, to maintain consistency, added link to `--android-device` one row
above now that we support anchor links in the docs.